Make architecture basic
