--- internal/terminal/handler/ghrepository/table.go
@@ -1,12 +1,0 @@
-package ghrepository
-
-import (
-	"github.com/charmbracelet/bubbles/table"
-)
-
-var tableColumnsGithubRepository = []table.Column{
-	{Title: "Repository", Width: 24},
-	{Title: "Default Branch", Width: 16},
-	{Title: "Stars", Width: 6},
-	{Title: "Workflows", Width: 9},
-}
