--- internal/terminal/handler/ghworkflow/table.go
@@ -1,10 +1,0 @@
-package ghworkflow
-
-import (
-	"github.com/charmbracelet/bubbles/table"
-)
-
-var tableColumnsWorkflow = []table.Column{
-	{Title: "Workflow", Width: 32},
-	{Title: "File", Width: 48},
-}
